task(data-explorer, components): move DefaultCardContent to components module
